move get_column_issues
